stars: moved key DB classes to separate files, error check for file writing
